Get kitchen-docker to install Chef
By default, test-kitchen is responsible for installing Chef.  This
involves downloading and installed it into the test container every
single time.  By having kitchen-docker do it instead, the result of
the install is cached making subsequent test runs considerably
shorter.

The down side of this approach is that newer versions of the Chef
client will not be used without manual intervention.