Revert "Only install .kitchen.local.yml when kitchen is enabled."
This reverts commit 18c9ab3e30402965ae6bc1282a47dd839e628eed.

When the test-kitchen gem is part of a cookbook's bundle, and the
included Rakefile loads kitchen's rake tasks, all calls to rake will
fail unless kitchen-vagrant and vagrant are both installed (as the
.kitchen.yml from the cookbook specifies the vagrant driver).  This is
true even when the kitchen tasks are not being called.

Always installing the local override ensures that the docker driver is
used (instead of the vagrant driver), and calls to rake succeed.